+++ src/parse_data.py
@@ -1,6 +1,6 @@
-# coding: utf-8
+# encoding: utf-8
 # author: yaoh.wu
-#
+
 import csv
 import logging
 # 原始数据合法起始字段正则
@@ -68,7 +68,8 @@ def simple_parse_data():
     """
     logging.info("simple parse data begin")
 
-    os.remove(parsed_data_file_path)
+    if os.path.exists(parsed_data_file_path):
+        os.remove(parsed_data_file_path)
 
     # 临时数据列表
     tmp_data_list = []
@@ -117,8 +118,11 @@ def divide_parsed_data():
     分为：用户访问数据 和 定时调度使用数据 两部分
     """
     logging.info("divide parsed data begin")
-    os.remove(user_data_file_path)
-    os.remove(schedule_data_file_path)
+
+    if os.path.exists(user_data_file_path):
+        os.remove(user_data_file_path)
+    if os.path.exists(schedule_data_file_path):
+        os.remove(schedule_data_file_path)
 
     # 用户数据临时数据列表
     user_tmp_list = []
@@ -156,10 +160,14 @@ def divide_parsed_data():
 
 def count_parsed_user_data():
     # 移除旧的存储文件
-    os.remove(tpl_count_file_path)
-    os.remove(user_count_file_path)
-    os.remove(user_role_count_file_path)
-    os.remove(common_count_file_path)
+    if os.path.exists(tpl_count_file_path):
+        os.remove(tpl_count_file_path)
+    if os.path.exists(user_count_file_path):
+        os.remove(user_count_file_path)
+    if os.path.exists(user_role_count_file_path):
+        os.remove(user_role_count_file_path)
+    if os.path.exists(common_count_file_path):
+        os.remove(common_count_file_path)
     """
     针对用户名称，模板名称，用户角色做一下计数
     """
